requirements.txt
fastapi
uvicorn
motor
pyjwt
python-dotenv
pydantic
python-multipart
